Revert "Revert "Update firefox gecko id""
This reverts commit cd607394a55facda82fd4e55f1b839222be85b3e.

The gecko id may be set to `arxiv-utils@j3soon.com` when uploading
the extension for the first time. If the gecko id isn't set in the
first upload, an id will be assigned and cannot be changed later.

In our case, the assigned gecko id is `ab779d78-7270-4ee8-9ee8-369d73508298`.

Commits related to this issue:
- cd607394a55facda82fd4e55f1b839222be85b3e
- 3e1c4daf3f9a577843c1ad31946eaec892c2c794
- 842ec362c0757f46fe4c08e0d710a118eb25c6ad
- 2d5ba9290c4a2c9df0ea2403a18836f823cba270
- dfe5c0d7424e8df40c8cbe4d66a5f8c76eee0916