Replace the `chrome` namespace with `browser` for Firefox
The asynchronous APIs in the `chrome` namespace only allow callbacks
(doesn't return a promise) in Manifest v2 (both in Chrome and Firefox).
This API is suboptimal and has since been changed to return a promise
in Manifest v3.

Currently, Chrome/Edge version of arxiv-utils is ported to Manifest v3,
while the Firefox version is still in Manifest v2. So we need to use
the `browser` namespace for Firefox so as to use `await` on the
asynchronous APIs (which return a promise).

The Firefox version can use the `chrome` namespace when ported to
Manifest v3.

This commit also fixes the issue in Firefox where the new tab will be opened at
the end of the tab list instead of next to the current tab.

References:
- https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/Chrome_incompatibilities#callbacks_and_promises
- https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/Chrome_incompatibilities#firefox_supports_both_the_chrome_and_browser_namespaces
- https://developer.chrome.com/docs/extensions/reference/tabs/#get-the-current-tab