Fix direct download by webextension download API
The new arxiv.org behavior redirects PDF to browser.arxiv.org, which
causes the original direct download using <a> links to fail due to the
violation of same-origin URL requirement. See:

    https://developer.mozilla.org/en-US/docs/Web/HTML/Element/a#attributes

To prevent similar issues from happening in the future, a more robust
way is to simply use the webextension download API along with send
message, so as to achieve real direct downloads.